Fix the tunnel's connect timeout and the target save's round trip
Socket.connect takes an Int timeout; the Long conversion was a compile
error the schema-dump workflow could not see (ksp never compiles) and
the first full CI build rejected. The save's round trip passed a
RemoteDesktopTarget to encodeRemoteDesktop, which takes the config -
restructured so the encode happens once and the decode is the guard it
was always meant to be.

+++ app/src/main/java/dev/eclipse/ssh/vnc/VncTunnel.kt
@@ -141,7 +141,7 @@ class VncTunnel(private val forwarding: PortForwardingManager) {
                 forward = handle
                 val socket = Socket()
                 socket.tcpNoDelay = true
-                socket.connect(InetSocketAddress(LOOPBACK, handle.boundPort), CONNECT_TIMEOUT_MS.toLong())
+                socket.connect(InetSocketAddress(LOOPBACK, handle.boundPort), CONNECT_TIMEOUT_MS)
                 this.socket = socket
                 val config = VernacularConfig().apply {
                     targetFramesPerSecond = framesPerSecond
